backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pymysql==1.1.0
cryptography==41.0.7
requests==2.31.0
aiofiles==23.2.1
python-dateutil==2.8.2
pydantic==2.5.0
pydantic-settings==2.1.0
dbutils==3.0.3
PyJWT==2.8.0
dashscope>=1.14.0
alibabacloud-outboundbot20191226>=1.0.0
alibabacloud-tea-openapi>=0.3.8
alibabacloud-tea-util>=0.3.11
